Move getbalances into_model functions into module
Move the into_model functions from mod.rs into the `into` module.  Code
move only, no other changes.
