+++ types/src/v19/wallet/into.rs
@@ -1,10 +1,44 @@
 // SPDX-License-Identifier: CC0-1.0
 
+use bitcoin::amount::ParseAmountError;
 use bitcoin::consensus::encode;
-use bitcoin::{BlockHash, SignedAmount, Transaction, Txid};
+use bitcoin::{Amount, BlockHash, SignedAmount, Transaction, Txid};
 
-use super::{GetTransaction, GetTransactionError};
+use super::{GetBalances, GetBalancesMine, GetBalancesWatchOnly, GetTransaction, GetTransactionError};
 use crate::model;
+
+impl GetBalances {
+    /// Converts version specific type to a version nonspecific, more strongly typed type.
+    pub fn into_model(self) -> Result<model::GetBalances, ParseAmountError> {
+        let mine = self.mine.into_model()?;
+        let watch_only = self.watch_only.map(|watch_only| watch_only.into_model()).transpose()?;
+
+        Ok(model::GetBalances { mine, watch_only })
+    }
+}
+
+impl GetBalancesMine {
+    /// Converts version specific type to a version nonspecific, more strongly typed type.
+    pub fn into_model(self) -> Result<model::GetBalancesMine, ParseAmountError> {
+        let trusted = Amount::from_btc(self.trusted)?;
+        let untrusted_pending = Amount::from_btc(self.untrusted_pending)?;
+        let immature = Amount::from_btc(self.immature)?;
+        let used = self.used.map(Amount::from_btc).transpose()?;
+
+        Ok(model::GetBalancesMine { trusted, untrusted_pending, immature, used })
+    }
+}
+
+impl GetBalancesWatchOnly {
+    /// Converts version specific type to a version nonspecific, more strongly typed type.
+    pub fn into_model(self) -> Result<model::GetBalancesWatchOnly, ParseAmountError> {
+        let trusted = Amount::from_btc(self.trusted)?;
+        let untrusted_pending = Amount::from_btc(self.untrusted_pending)?;
+        let immature = Amount::from_btc(self.immature)?;
+
+        Ok(model::GetBalancesWatchOnly { trusted, untrusted_pending, immature })
+    }
+}
 
 impl GetTransaction {
     /// Converts version specific type to a version nonspecific, more strongly typed type.
