+++ types/src/v19/wallet/mod.rs
@@ -6,12 +6,10 @@
 
 mod into;
 
-use bitcoin::amount::ParseAmountError;
-use bitcoin::{Amount, Transaction};
+use bitcoin::Transaction;
 use serde::{Deserialize, Serialize};
 
 use super::{Bip125Replaceable, GetTransactionDetail, GetTransactionError};
-use crate::model;
 
 /// Result of the JSON-RPC method `getbalances`.
 ///
@@ -50,39 +48,6 @@ pub struct GetBalancesWatchOnly {
     pub untrusted_pending: f64,
     /// Balance from immature coinbase outputs.
     pub immature: f64,
-}
-
-impl GetBalances {
-    /// Converts version specific type to a version nonspecific, more strongly typed type.
-    pub fn into_model(self) -> Result<model::GetBalances, ParseAmountError> {
-        let mine = self.mine.into_model()?;
-        let watch_only = self.watch_only.map(|watch_only| watch_only.into_model()).transpose()?;
-
-        Ok(model::GetBalances { mine, watch_only })
-    }
-}
-
-impl GetBalancesMine {
-    /// Converts version specific type to a version nonspecific, more strongly typed type.
-    pub fn into_model(self) -> Result<model::GetBalancesMine, ParseAmountError> {
-        let trusted = Amount::from_btc(self.trusted)?;
-        let untrusted_pending = Amount::from_btc(self.untrusted_pending)?;
-        let immature = Amount::from_btc(self.immature)?;
-        let used = self.used.map(Amount::from_btc).transpose()?;
-
-        Ok(model::GetBalancesMine { trusted, untrusted_pending, immature, used })
-    }
-}
-
-impl GetBalancesWatchOnly {
-    /// Converts version specific type to a version nonspecific, more strongly typed type.
-    pub fn into_model(self) -> Result<model::GetBalancesWatchOnly, ParseAmountError> {
-        let trusted = Amount::from_btc(self.trusted)?;
-        let untrusted_pending = Amount::from_btc(self.untrusted_pending)?;
-        let immature = Amount::from_btc(self.immature)?;
-
-        Ok(model::GetBalancesWatchOnly { trusted, untrusted_pending, immature })
-    }
 }
 
 /// Result of the JSON-RPC method `gettransaction`.
